requirements.txt
# Telegram Client API
telethon==1.34.0

# Database
aiosqlite==0.19.0

# Scheduler
apscheduler==3.10.4

# Web interface
flask==3.0.0
flask-cors==4.0.0

# Additional dependencies
python-dotenv==1.0.0
requests==2.31.0
